requirements.txt
sklearn
pandas
seaborn
dvc
google-api-python-client
dask[dataframe]
gcsfs
google-cloud
google-cloud-storage
joblib
xgboost
tensorflow
